pyproject.toml
[project]
name = "matsuzawa-ac-cot"
version = "0.1.0"
description = "Abstract-then-Compact Chain-of-Thought experiments on GSM8K"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "together>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
